requirements.txt
Flask==0.12.2
flask_oauthlib==0.9.4
python-dotenv==0.7.1
